requirements.txt
flask
flask-sqlalchemy
flask-login
flask-bcrypt
flask-wtf
python-dateutil